Modified enqueue of script and styles
When initially deploying the alpha color picker, the CSS and JS URIs are hard-coded as those the original author uses (/admin/customizer/alpha-color-picker/alpha-color-picker.js) . Other developers may wish to locate the alpha color picker in another folder in the theme (e.g. in inc folder)

The added code works out what the URI of the folder the php file is located in, and therefore works out the correct URI of the CSS and JS files without requiring a developer to edit the source code of the alpha color picker, allowing for easier updating without having to edit the URI everytime.